LPD-31556 Moves more methods to Base class
