+++ modules/apps/fragment/fragment-renderer-categorization-inputs-impl/bnd.bnd
@@ -1,3 +1,4 @@
 Bundle-Name: Liferay Fragment Renderer Categorization Inputs Implementation
 Bundle-SymbolicName: com.liferay.fragment.renderer.categorization.inputs.impl
 Bundle-Version: 1.0.9
+-dsannotations-options: inherit
