+++ modules/apps/fragment/fragment-renderer-categorization-inputs-impl/src/main/java/com/liferay/fragment/renderer/categorization/inputs/internal/CategoriesInputFragmentRenderer.java
@@ -7,32 +7,19 @@ package com.liferay.fragment.renderer.categorization.inputs.internal;
 
 import com.liferay.asset.kernel.model.AssetVocabularyConstants;
 import com.liferay.asset.taglib.servlet.taglib.AssetCategoriesSelectorTag;
-import com.liferay.fragment.constants.FragmentConstants;
 import com.liferay.fragment.model.FragmentEntryLink;
 import com.liferay.fragment.renderer.FragmentRenderer;
 import com.liferay.fragment.renderer.FragmentRendererContext;
 import com.liferay.fragment.util.configuration.FragmentEntryConfigurationParser;
-import com.liferay.frontend.taglib.clay.servlet.taglib.AlertTag;
 import com.liferay.info.constants.InfoItemScopeConstants;
 import com.liferay.info.item.InfoItemServiceRegistry;
 import com.liferay.info.item.provider.InfoItemCategorizationProvider;
 import com.liferay.info.item.provider.InfoItemScopeProvider;
-import com.liferay.layout.constants.LayoutWebKeys;
-import com.liferay.layout.page.template.model.LayoutPageTemplateStructure;
-import com.liferay.layout.page.template.service.LayoutPageTemplateStructureLocalService;
-import com.liferay.layout.util.constants.LayoutDataItemTypeConstants;
 import com.liferay.layout.util.structure.FormStyledLayoutStructureItem;
-import com.liferay.layout.util.structure.FragmentStyledLayoutStructureItem;
-import com.liferay.layout.util.structure.LayoutStructure;
-import com.liferay.layout.util.structure.LayoutStructureItem;
-import com.liferay.layout.util.structure.LayoutStructureItemUtil;
 import com.liferay.petra.string.StringPool;
-import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.json.JSONException;
 import com.liferay.portal.kernel.json.JSONFactory;
 import com.liferay.portal.kernel.json.JSONObject;
-import com.liferay.portal.kernel.json.JSONUtil;
-import com.liferay.portal.kernel.language.Language;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
@@ -60,11 +47,6 @@ import org.osgi.service.component.annotations.Reference;
  */
 @Component(service = FragmentRenderer.class)
 public class CategoriesInputFragmentRenderer extends BaseInputFragmentRenderer {
-
-	@Override
-	public String getCollectionKey() {
-		return "INPUTS";
-	}
 
 	@Override
 	public String getConfiguration(
@@ -102,19 +84,7 @@ public class CategoriesInputFragmentRenderer extends BaseInputFragmentRenderer {
 		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
 			"content.Language", getClass());
 
-		return _language.get(resourceBundle, "categories");
-	}
-
-	@Override
-	public int getType() {
-		return FragmentConstants.TYPE_INPUT;
-	}
-
-	@Override
-	public String getTypeOptions() {
-		return JSONUtil.put(
-			"fieldTypes", JSONUtil.putAll("categorization")
-		).toString();
+		return language.get(resourceBundle, "categories");
 	}
 
 	@Override
@@ -125,7 +95,7 @@ public class CategoriesInputFragmentRenderer extends BaseInputFragmentRenderer {
 
 		try {
 			FormStyledLayoutStructureItem formStyledLayoutStructureItem =
-				_getFormStyledLayoutStructureItem(
+				getFormStyledLayoutStructureItem(
 					fragmentRendererContext.getFragmentEntryLink(),
 					httpServletRequest);
 
@@ -151,7 +121,7 @@ public class CategoriesInputFragmentRenderer extends BaseInputFragmentRenderer {
 			PrintWriter printWriter = httpServletResponse.getWriter();
 
 			if (!infoItemCategorizationProvider.supportsCategorization()) {
-				_writeDisabledCategorizationAlert(
+				writeDisabledCategorizationAlert(
 					fragmentRendererContext, httpServletRequest,
 					httpServletResponse, printWriter);
 
@@ -207,62 +177,6 @@ public class CategoriesInputFragmentRenderer extends BaseInputFragmentRenderer {
 		}
 	}
 
-	private FormStyledLayoutStructureItem _getFormStyledLayoutStructureItem(
-			FragmentEntryLink fragmentEntryLink,
-			HttpServletRequest httpServletRequest)
-		throws PortalException {
-
-		LayoutStructure layoutStructure = _getLayoutStructure(
-			fragmentEntryLink, httpServletRequest);
-
-		FragmentStyledLayoutStructureItem fragmentStyledLayoutStructureItem =
-			(FragmentStyledLayoutStructureItem)
-				layoutStructure.getLayoutStructureItemByFragmentEntryLinkId(
-					fragmentEntryLink.getFragmentEntryLinkId());
-
-		if (fragmentStyledLayoutStructureItem == null) {
-			return null;
-		}
-
-		LayoutStructureItem layoutStructureItem =
-			LayoutStructureItemUtil.getAncestor(
-				fragmentStyledLayoutStructureItem.getItemId(),
-				LayoutDataItemTypeConstants.TYPE_FORM, layoutStructure);
-
-		if (!(layoutStructureItem instanceof FormStyledLayoutStructureItem)) {
-			return null;
-		}
-
-		return (FormStyledLayoutStructureItem)layoutStructureItem;
-	}
-
-	private LayoutStructure _getLayoutStructure(
-			FragmentEntryLink fragmentEntryLink,
-			HttpServletRequest httpServletRequest)
-		throws PortalException {
-
-		LayoutStructure layoutStructure = null;
-
-		if (httpServletRequest != null) {
-			layoutStructure = (LayoutStructure)httpServletRequest.getAttribute(
-				LayoutWebKeys.LAYOUT_STRUCTURE);
-		}
-
-		if (layoutStructure == null) {
-			LayoutPageTemplateStructure layoutPageTemplateStructure =
-				_layoutPageTemplateStructureLocalService.
-					fetchLayoutPageTemplateStructure(
-						fragmentEntryLink.getGroupId(),
-						fragmentEntryLink.getPlid());
-
-			layoutStructure = LayoutStructure.of(
-				layoutPageTemplateStructure.getData(
-					fragmentEntryLink.getSegmentsExperienceId()));
-		}
-
-		return layoutStructure;
-	}
-
 	private int[] _getVisibilityTypes(FragmentEntryLink fragmentEntryLink) {
 		String vocabularyType = GetterUtil.getString(
 			_fragmentEntryConfigurationParser.getFieldValue(
@@ -283,29 +197,6 @@ public class CategoriesInputFragmentRenderer extends BaseInputFragmentRenderer {
 		return new int[] {AssetVocabularyConstants.VISIBILITY_TYPE_PUBLIC};
 	}
 
-	private void _writeDisabledCategorizationAlert(
-			FragmentRendererContext fragmentRendererContext,
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse, PrintWriter printWriter)
-		throws Exception {
-
-		if (!fragmentRendererContext.isEditMode()) {
-			return;
-		}
-
-		AlertTag alertTag = new AlertTag();
-
-		alertTag.setMessage(
-			_language.get(
-				fragmentRendererContext.getLocale(),
-				"categorization-is-disabled-for-the-selected-content"));
-		alertTag.setTitle(
-			_language.get(fragmentRendererContext.getLocale(), "info"));
-
-		printWriter.write(
-			alertTag.doTagAsString(httpServletRequest, httpServletResponse));
-	}
-
 	private static final Log _log = LogFactoryUtil.getLog(
 		CategoriesInputFragmentRenderer.class);
 
@@ -317,12 +208,5 @@ public class CategoriesInputFragmentRenderer extends BaseInputFragmentRenderer {
 
 	@Reference
 	private JSONFactory _jsonFactory;
-
-	@Reference
-	private Language _language;
-
-	@Reference
-	private LayoutPageTemplateStructureLocalService
-		_layoutPageTemplateStructureLocalService;
 
 }
